+++ Changelog.md
@@ -14,9 +14,17 @@ Bug fixes
 
 Unstable features
 
+* `unstable-client-transport`: Fixed an issue with slow responses in the
+  `multi_stream` transport by not waiting in the first iteration if an
+  underlying stream reports its connection being closed. ([#338])
+* New unstable feature `unstable-validator` that adds a DNSSEC validator.
+  ([#328])
+
 Other changes
 
+[#328]: https://github.com/NLnetLabs/domain/pull/328
 [#333]: https://github.com/NLnetLabs/domain/pull/333
+[#338]: https://github.com/NLnetLabs/domain/pull/338
 
 ## 0.10.1